examples/discord/dspy-personas/pyproject.toml
[tool.poetry]
name = "honcho-dspy-personas"
version = "0.1.0"
description = ""
authors = ["Vineeth Voruganti <13438633+VVoruganti@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
honcho-ai = {path = "../../..", develop = true}
langchain = "^0.2.0"
langchain-openai = "^0.1.7"
dspy-ai = "^2.4.9"
python-dotenv = "^1.0.1"
py-cord = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
